tests: refactored many module names
