+++ tests/updatehub/test_updatehub.py
@@ -5,7 +5,7 @@ import unittest
 from unittest.mock import patch
 
 from uhu.core.updatehub import (
-    finish_package, ObjectUploadResult, push_package,
+    finish_package, ObjectUploadResult, push_package, get_package_status,
     upload_metadata, upload_object, upload_objects, UpdateHubError)
 from uhu.http import HTTPError
 
@@ -120,3 +120,24 @@ class FinishPackageTestCase(unittest.TestCase):
     def test_raises_error_when_fail(self, http):
         with self.assertRaises(UpdateHubError):
             finish_package('1234')
+
+
+class PackageStatusTestCase(unittest.TestCase):
+
+    @patch('uhu.core.updatehub.http.get')
+    def test_returns_status_when_success(self, mock):
+        mock.return_value.ok = True
+        mock.return_value.status_code = 200
+        mock.return_value.json.return_value = {'status': 'done'}
+        status = get_package_status('1234')
+        self.assertEqual(status, 'done')
+
+    @patch('uhu.core.updatehub.http.get')
+    def test_raises_error_if_server_error(self, mock):
+        effects = [{}, ValueError]
+        mock.return_value.ok = False
+        mock.return_value.status_code = 404
+        mock.return_value.json.side_effect = effects
+        for _ in effects:
+            with self.assertRaises(UpdateHubError):
+                get_package_status('1234')
